# AsyncLogger: enqueue logging effect instead of a log message