My attempt to fix https://github.com/valskalla/odin/issues/460.

This approach fixes the underlying issue, but it affects the performance as well.

### 1) AsyncLoggerBenchmark

Original benchmarks of async logger are a little bit of cheating. Since they are basically measuring the `queue.offer` operation. 

**Before (master branch):**
```
Benchmark                             Mode  Cnt    Score     Error  Units
AsyncLoggerBenchmark.msg              avgt   25  1241.937 ±  276.487  ns/op
AsyncLoggerBenchmark.msgAndCtx        avgt   25  1274.021 ±   82.916  ns/op
AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  2021.246 ± 1326.558  ns/op
```

**After (fix branch):**
```
Benchmark                             Mode  Cnt    Score     Error  Units
AsyncLoggerBenchmark.msg              avgt   25  699.431 ±  73.839  ns/op
AsyncLoggerBenchmark.msgAndCtx        avgt   25  716.672 ± 105.136  ns/op
AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  926.306 ± 568.412  ns/op
```

### 2) Measure the cost of the `drain` operation

I decided to measure the cost of the `drain` operation as well. I tweaked `AsyncLoggerBenchmark` for this purpose:
```diff
val (asyncLogger: Logger[IO], cancelToken: IO[Unit]) =
  fileLogger[IO](fileName).withAsync(
+   timeWindow = Int.MaxValue.seconds, // to prevent background flushes
    maxBufferSize = Some(1000000)
  ).allocated.unsafeRunSync()
    
@Benchmark
@OperationsPerInvocation(1000)
def msg(): Unit = 
  (1 to 1000).toList.traverse(_ => asyncLogger.info(message))
+   .flatMap(_ => asyncLogger.drain) 
    .unsafeRunSync()
```

**Before (master branch):**
```
Benchmark                             Mode  Cnt     Score    Error  Units
AsyncLoggerBenchmark.msg              avgt   25  1021.347 ± 32.193  ns/op
AsyncLoggerBenchmark.msgAndCtx        avgt   25  1013.955 ± 68.780  ns/op
AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  4804.478 ± 99.426  ns/op
```

**After (fix branch):**
```
Benchmark                             Mode  Cnt     Score     Error  Units
AsyncLoggerBenchmark.msg              avgt   25  3483.380 ± 125.855  ns/op
AsyncLoggerBenchmark.msgAndCtx        avgt   25  3478.960 ±  95.501  ns/op
AsyncLoggerBenchmark.msgCtxThrowable  avgt   25  7026.014 ± 107.790  ns/op
```